Criacao da API  inicial para teste de FindALL planos Rasmoo-PLUS

+++ ws-rasmoo-plus/src/main/java/com/client/ws/rasmooplus/model/SubscriptionsType.java
@@ -22,7 +22,7 @@ public class SubscriptionsType implements Serializable {
     private String name;
 
     @Column(name = "access_months")
-    private int accessMonths;
+    private Long accessMonths;
 
     private Double price;
